# docs/swagger.yaml
definitions:
...
  handler.Error:
    properties:
      message:
        type: string
    type: object
...
  handler.ResponseError:
    properties:
      errors:
        items:
          $ref: '#/definitions/handler.Error'
        type: array
    type: object
  handler.ResponseSuccess:
    properties:
      data: {}
    type: object
...
host: localhost:3000
info:
  contact:
    email: romain.rtestard@gmail.com
    name: Romain Testard
  description: Server allowing interaction with Seller Payout Domain
  title: SellerPayout Rest Server
  version: "1.0"
paths:
...
  /payouts/:seller_id:
    get:
      consumes:
      - application/json
      description: Create Seller.
      parameters:
      - description: Seller ID query parameter
        in: query
        name: seller_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ResponseError'
      summary: Endpoint to retrieve payouts for a specific seller.
      tags:
      - Seller
...
swagger: "2.0"